第二次更改 增加了service接口,重新生成了数据库 因为数据库表格有人建的太差 我重新建了dept,emp,job表格 蔡文翠由于表格有注释,也设置了默认值 故没有改动 前端测试能登陆成功 有list,add,update,delete等功能 切记 此次项目删除是改status的状态 由于我和王姣分到了同一个模块 故可能有代码冲突的可能 请善用Javadoc注释 代码太多真的看不到 请尽量类名,方法名能表达你的这个方法，这个类名的作用 请尽量在方法名使用javadoc注释 描述这个方法是如何起作用，有什么有作用

+++ tabac/src/main/java/com/xr/mapper/SysDeptMapper.java
@@ -12,15 +12,13 @@ public interface SysDeptMapper {
 
     /**
      * delete by primary key
-     *
      * @param id primaryKey
      * @return deleteCount
      */
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     /**
      * insert record to table
-     *
      * @param record the record
      * @return insert count
      */
@@ -28,7 +26,6 @@ public interface SysDeptMapper {
 
     /**
      * insert record to table selective
-     *
      * @param record the record
      * @return insert count
      */
@@ -38,11 +35,10 @@ public interface SysDeptMapper {
 
     /**
      * select by primary key
-     *
      * @param id primary key
      * @return object by primary key
      */
-    SysDept selectByPrimaryKey(Long id);
+    SysDept selectByPrimaryKey(Integer id);
 
     int updateByExampleSelective(@Param("record") SysDept record, @Param("example") SysDeptExample example);
 
@@ -50,7 +46,6 @@ public interface SysDeptMapper {
 
     /**
      * update record selective
-     *
      * @param record the updated record
      * @return update count
      */
@@ -58,7 +53,6 @@ public interface SysDeptMapper {
 
     /**
      * update record
-     *
      * @param record the updated record
      * @return update count
      */
